Web: fix selection of UOTD
The default for uotd_time in the profile table is NULL but new profiles created after commit 6543928 have uotd_time=0. This means they are not eligible for UOTD because the default candidate queries only check if uotd_time is null. This fixes that and considers NULL and 0 as 'profile has not been UOTD but can be a candidate'.