Build: remove --disable-wx-debug
WxWidgets switched to using -DNDEBUG to disable debugging in version 2.9.1 and enabling it by default. My previous change (530eb18) made the debug version the default and one had to disable the debugging for release builds. This is fixed now and the non-debug version is the default. Now WxWidgets debug assertions are switched on if --enable-debug is specified.

fixes #1476