lib, Unix: remove chown/chmod stuff from boinc_copy().
boinc_copy() is used in both client and server.
The chown/chmod was added Jan 6 2009,
apparently to deal with a client problem.
More recently, error checking was added to the chown/chmod calls.
On the server (e.g. sample assimilator) these fail,
and boinc_copy() returns error (even though the file was actually copied).

1) let's not overload functions.  Functions in general should do 1 thing.
To this end, I moved the chown/chmod to a new function, boinc_copy_attributes().

2) It's not clear to me that we need the chown/chmod stuff,
either on client or server.
On the Unix client, we already have FILE_INFO::set_permissions(),
which sets the rwx bits of files copied into slot directories, as appropriate.
The chown() will be a no-op, since the client creates
both the original file (in project dir) and the copy.

So I'm not going to call boinc_copy_attributes() in the client.
Let's see if any problems arise in Unix with apps that
used <copy_file> for executables.