MGR: Bring up the attach wizard project welcome page if the project url has been tampered with by the installer.
If the project URL is embedded in the installer via project_init.xml (and presumably code-signed by somebody) we can allow the software to automatically attach to that project with a setup cookie or authenticator.  If somebody attempts to use that installer for another project, we need to make sure the volunteer knows which project the client is going to attempt to attach to.