Drupal: Fixed display logic in forum post template
Part of addressing first post pinning in 94b02e95 involved hiding the topic starter content on all but the last page when sort order is set to newest post first. The logic was incomplete, however, and was hiding the topic starter post content on the first page even in cases where it is only one page.

https://dev.gridrepublic.org/browse/DBOINCP-28