VBOX: Take a screen shot on failure and dump the base64 encoded image to stderr.
Report the MD5 signature in the stderr output as well so that duplicate screen shots can be quickly identified.  Of course this assumes that the exact same PNG is generated across different machine types and video cards.  Currently I believe that it should given the hardware in question is virtualized.