VBOX: Treat ERR_INVALID_PARAM from create_vm() as an unrecoverable error.
Since most everyone is running something more recent that BOINC 7.2.16, we can assume that the hardware just does not support hardware acceleration.  Inform BOINC that the VM Extensions have been disabled so that the server can either assign a 32-bit VM which can use software emulation, or just not assign future work to the host.