Client: add OS version detection to test tool
Trying to read the output of lsb_release -a first and then the content of /etc/os-release to determine the distribution name and version to add to os_name and os_version. This will not cover all distributions but the major ones at least adopted the /etc/os-release file providing minimal information.