client: fix problems w/ tempfile generation for async file operations
Turns out tempnam() can't be used;
it gives priority to the env var TMPDIR in deciding where to
put the temp file (overriding the prefix arg)
and this could be on a different volume.
Instead, use msktemp (Unix) and GetTempFileName (Win).
And factor this out into a util function, boinc_temp_file().