remote job submission: improve error handling.
The way errors were being handled in submit_rpc_handler.php,
the replies could end up looking like:
<error>
    ...
    </error>
<actual_reply>
    ...
</actual_reply>

which doesn't parse with PHP's simplexml because there's no outer element.
So the PHP interface to job submission didn't work in some cases.

To fix this, I changed the RPC replies to have an enclosing element,
which is the name of the RPC, e.g.
<submit_batch>
    <error>
        ...
    </error>
    ... other stuff
</submit_batch>

RPC replies should now always be valid XML, errors or not.