BOINC GAHP: implement 'local input template' feature
If an input file name ends in "_input_tempplate.xml", it is treated as an input
template file for that batch. It is uploaded to the server as
"templates/batch_<batch_id>_in" and used for the currently submitted batch.

Once the GAHP protocol is extended to specify these files separately from
the ordinary input files, only the process_input_template() function
needs to be changed.