Drupal: update preference wording
The wording of preferences now mostly match those used on the regular BOINC website so volunteers are not confused about the meanings.
Special note: The preferences run_on_batteries, run_if_user_active and run_gpu_if_user_active are now shown invers to how they are stored because of the changed description.