client: fix bug that could cause idle CPUs/GPUs.
Review: job scheduling has 2 phases:
1) Make a list of jobs to run.  Add enough jobs to use all resources.
2) Actually run the jobs and preempt existing jobs.

The problem: checking for RAM usage limits
(i.e. making sure the sum of working sets is < RAM usage prefs)
is done in 2) but not 1).
So on a 1 CPU machine we might make a run list consisting of a single job,
which turns out not to fit in available RAM,
and we end up running nothing.

Solution: when we add a job to the run list that previously
exceeded RAM limits, don't count its resource usage.
That way we'll add more jobs to the run list,
and we'll have something to run in the end.