Client: reorder host.os_version for Linux
Move the user relevant information (Linux Distribution) to the beginning of the string and put the more technical information (kernel version) in square brackets at the end. This may break parsing in a custom scheduler if someone expects the kernel version to be the first string of the field on Linux hosts.