Client: unescape special characters read from /etc/os-release
The information in /etc/os-release is allowed to contain some special shell characters that are escaped so the file can be easily read by shell scripts. We store this information in the database and do our own escaping. The strip_quotes() function now does not remove quotes from the end of the string if they are escaped.