Validator: fix path and filename when in standalone mode
When in standalone mode (test framework) the name is actually the path supplied by the user. This will extract the filename from the given path and fill in the correct values of the OUTPUT_FILE_INFO object.
This is important so that the test framework is consistent with the actual behaviour and projects don't need to put in extra code to distinguish between test mode and normal mode.