Drupal:
Fixed bug where users who chose forum preference "Newest Post First" had an error in how a new forum post was previewed. Added a if conditional where if the number of comments is zero, the forum topic template does not try to create an additional set of div-tags. As a result the Preview link now works.

https://dev.gridrepublic.org/browse/DBOINCP-300