Drupal: convert preset attribute to tag
The additional preset attribute in the venue tag is breaking preference parsing in the scheduler. We now use a tag instead of an attribute and clear any old attributes.
This is only a problem with generalprefs because they get propagated to other projects that do not have a scheduler which ignores the attribute. The Einstein@Home scheduler (only project to use the Drupal webcode right now) was patched to handle extra attributes so we don't need to strip the attribute from projectprefs.