Locale: update tool to update translations
This finishes the transition from the txgh bridge to using the transifex client to download translations and update them in the github repository. A BOINC developer needs to have write access to the github repository and must be a maintainer of the transifex project.
Don't forget to initialize the transifex client with the transifex credentials (http://docs.transifex.com/client/init/) before first use of the script.