Locale: tweak filepaths in templates
Omitting the first directory in the filename of translatable strings is ambiguous if you don't know what directory this template was build from. So this adds back the top level directory and the filename in comments inside the pot files is now always relative to the source directory.
I also added the generic language name header to the BONIC-Web ressource becaue it also needs to know the language names and not necessarily gets the info from BOINC-Project-Generic.