vboxwrapper: add <copy_cmdline_to_shared> option to job file
Tells vboxwrapper to copy its cmdline args to shared/cmdline.

This lets you use the cmdline mechanism as a way to pass
info to the VM that would otherwise have to go in a file.

Note: the cmdline can be fairly large.
It shares the wu.xml_doc field with file descriptions.
This field can be up to 64KB (MySQL blob size)