Manager: fix select computer dialog for localhost
The hostname "localhost" was recognized as a special name which always prompted the Manager to try to read contents of gui_rpc_auth.cfg. The Manager used an empty password if the file is not readable but then can't connect because the Client actually uses a password. Any user supplied password was overwritten which made the dialog not work when using "localhost" but work when used with "127.0.0.1".
Now the dialog checks if the user has entered an empty password which prompts reading gui_rpc_auth.cfg and uses the user supplied password otherwise.

Fixes #1306