remote file management: allow long filenames
Originally, the idea for remote file management was that
the physical name of a file was its md5, or jf_md5.
This was changed to let the submitter choose the name.
But in the DB (job_file.md5) we only kept 64 chars.

Change the field name to job_file.name, and change it to varchar(255)