BOINC web: fix bug in get_platforms()
There was a logic error that would cause every load of
the project page to initiate a web request to GPUgrid.
When GPUgrid was down, this could create lots of sockets,
which eventually would prevent incoming connections.
I think this is the reason for the sporadic slowdowns
of the BOINC web site.