Mac replace all Mac APIs deprecated in OS 10.10, except in MacAccessibility.mm. Updating the accessibility support will require a complete rewrite to use Apple's new NSAccessibility APIs, which are not available on systems prior to OS 10.10, while retaining the current code for use with older versions of OS X. This will require far more time and effort than I can spend at this time.
Note also that Apple has deprecated the UNIX standard call system(const char *) as ofd OS 10.10 and says to use posix_spawn() calls instead. I have added a macro to the precompiled header MacGUI.pch for the Manager builds that converts the calls automatically, but have not been able to do so for the other BOINC modules. If future code in the client, libraries, etc. calls system(), this will have to be handled on a case by case basis, as I have done in boinc_rename_aux() in lib/filesys.cpp and CLIENT_STATE::write_state_file() in client/cs_statefile.cpp.