Build Mac: enhanced dependency build scripts
All scripts modified to use a prefix and install libraries into a new directory. This saves space on CI systems and should simplify the Xcode project a bit. Some functionality, like debug builds of wxWidgets is not implemented yet (They are too big for CI systems to cache).

This is an intermediate commit to test building of dependencies on Travis-CI. The XCode project needs to be adjusted for the new include directory in order to build BOINC components.