Drupal: implement creditonly switch for team_email_list.php RPC
This is used by stats sites to regularly query user credit stats during a special event. The default implementation is inefficient as it creates a DB query per team member which is slowing down the query in big teams. The creditonly switch is also available in upstream where it supresses certain xml outputs. We still don't support all the fields that upstream supports but the missing fields are not credit relevant.