Drupal: Fixed bug where navigation bar menu on desktop site was too wide due to translated text.
Menu bar text translations may replace English words with ones which are longer due to differences in language. The result is a navigation menu bar that is larger than 980px, which is the desktop fixed layout width.
JQuery code added to detect these oversize menus and display a hamburger mobile menu instead. Adjusted the CSS as well to handle this new layout.