Build: use default sysconfdir instead of /etc
When using --prefix this override tries to install files from client/scripts to /etc instead of ${prefix}/etc. This is against what the user wants and leads to an error message as normally only root can install files to /etc. There is no need to be backwards compatible in this place as installs are usually done via repository packages.
This still uses /etc to determine where the config file should be copied to which is ok because the user most likely wants to use the Client on the machine it was built on.

Note: The startup scripts and config files in client/scripts are outdated and might not work on recent Linux distributions. Newer ones can be found in the Debian/Fedora packages.