SCHED_BATCH -> SCHED_IDLE
It was a BOINC RasPi forum entry that let me revisit chrt and scheduler policies+priorities. The tasks are executed (if !highpriority) with the SCHED_BATCH policy. This also let me unearth again that SCHED_IDLE is what in the very early UNIX days was reached by renicing to the highest values but today SCHED_IDLE is less of a priority than SCHED_BATCH with nicelevel 19.  I just executed the "chrt -i ..." manually and it seems just fine. Graphics still work. Communication with the BOINC servers is performed with a higher priority, which lets me think that we should expect any problem. So I suggest to have SCHED_IDLE the default for compute jobs.
Best,
Steffen