Mac installer & screensaver: Allow building on newer versions of Xcode while preserving compatibility back to OS 10.6.
Problem: OS 10.6 and OS 10.7 require screensavers to be built with Garbage Collection (GC) enabled, but Xcode versions after 5.0.2 do not support GC.

Solution: save a compressed copy of the screensaver executable built with GC on GIT as BOINCSaver_MacOS10_6_7.zip, and update the Xcode project to build the screensaver with Automatic Reference Counting (ARC). Modify the Mac installer build script to add a copy of BOINCSaver_MacOS10_6_7 to the BOINCSaver bundle, so the installer installs the screen saver with both versions of the executable. Modify PostInstall.app to select the appropriate executable for the version of OS X at install time, rename it if necessary and delete the other one.