Adding user->id to be return with an e-mail lookup
For 3rd parties e-mail verification is the only way to verify users. This allows third parties to match a users id by his e-mail address, and thus securely identify their ownership of that account. 

Using this information we can access show_user.php?userid=id and be certain of an e-mails credit ownership.