Web: fix sample project_specific prefs
The allow_non_preferred_apps tag was renamed in 671446ced1 for the scheduler but the sample project_specific prefs were not updated.

Existing projects need to update the file on their own if they use a scheduler that was build after 671446ced1 and rename the old tag in the DB. See html/ops/fix_project_prefs.php to do that.