web: team-related fixes
At some point we added several team categories (national, computer type, etc.)
but there were a couple of places in the code where the # of categories
was hardwired, so the new categories weren't shown.

Also, for white-backbround color schemes,
in table headers with bg-default,
links were invisible because the text color was same as background.
Fix this in custom.css.