FUH: reopen logfile on SIGUSR2
This is needed in FCGI mode where the logfile is opened only at startup. The logfile is usually renamed and the still running process is still writing to it. Sending a SIGUSR2 to the process will reopen the original logfile.
The FastCGI spec lists SIGUSR1 as "gracefull exit" and the latest version of multiwatch sends it as a termination signal.

Also hide location of config file on server from user and flush FCGI log before exiting on a signal.