mgr: remove CDlgEventLogListCtrl::GetBackgroundColour()
Since the method is overridden to be private it isn't callable from outside the class and the method isn't used inside the class either.

And overriding a base class method just to return a constant color isn't very useful when the same effect can be had by, e.g., calling SetBackgroundColour() in class' constructor.

On the other hand, trying to force the background color to white may conflict with the color scheme user has chosen and should not be done.

All in all, remove the code as not useful.

Part of #1083.