mgr: adjust Event Log stripe color
On Windows, wxListCtrl has a bug that makes it draw selected unfocused rows with colors used for selected focused rows when the row background color is wxSYS_COLOUR_BTNFACE.

With default color schemes, on Windows XP the list is drawn correctly, but on Windows 10 the (240,240,240) used in stripe color is coincidentally the same as wxSYS_COLOUR_BTNFACE and selected rows are drawn with wrong colors.

Fix this by using the same stripe color that wxWidgets would use by default and that is used in Projects, Tasks and Transfers lists. Future proof the code by including a workaround just in case Windows changes color scheme again. If such color scheme change occurs then the same workaround needs to be added to Projects, Tasks and Transfers lists.

Fixes #1083.