Add generated systemd unit
Both Fedora and Debian have their own systemd units for boinc;
this is based on elements of both so we stop duplicating effort.

This also adds minimal confinement to protect the home directories.

Also added clean to init.d script and made it only install if a
init.d directory already exists.

Closes: #2255