Remove HOST_USAGE.max_ncpus.  It's not used anywhere.
avg_ncpus is always used instead.

Note: there are two other things named "max_ncpus".
One is a computing prefs field (limit on # of CPUs to use).
The other, a field of SCHED_CONFIG,
is the max # of CPUs the scheduler will believe a client has
in deciding how many jobs to send it.