web: Exiting users agree to terms of use when logging in.
Renamed 'optin consent' to 'agree to terms of use'. New config.xml variable name.
Decoupled new user agree to terms of use from existing users. If there is a ENROLL terms of use, and the project has enabled the 'enable_login_mustagree_termsofuse' boolean, then existing users must agree to the terms of use when the login, and this is recorded in the consent table.