Fix eol to be by default LF
In case that boinc is cloned on Windows, git adds CRLF to files that were not explicitly specified. When those files were copied into a Linux directly, without committing them to git, the line endings were preserved, and compilation failed.

It would be much more convenient to have all files cloned with expected line endings to overcome this issue.

This issue may occur when the development is done on Windows and build on Linux using Docker for example.