lib: fix wide char conversion functions
When MultiByteToWideChar and WideCharToMultiByte are told to process entire input strings the functions include null terminators in required buffer sizes. std::string on the other hand stores embedded null characters in the string it manages if it's requested to do so.

The way the code constructs std::string's from wide char function outputs makes null terminators part of the strings. This breaks concatenating strings together.

Fix the code by not copying the last character, i.e. the null terminator, to std::string.