scheduler: add support for <physical_cpus> and <max_threads>N [M]</max_threads> elements in XML plan class spec.
Change variables names from effective_ncpus to usable_ncpus_*

Note 1: pre-7.14 clients, and clients not on Win/Mac/Linux,
don't report # of physical CPUs.
In this case we set it to (# logical CPUs)/2.
This is necessary to ensure that VM jobs will run on such hosts
if they're hyperthreaded.

Note 2: The # of threads used by the app (passed to it as --nthreads)
and HOST_USAGE::avg_ncpus (with is in terms of logical CPUs)
are now separate entities.
On a hyperthreaded machine, a job with <physical_cpus> that
uses 4 threads will be shown (in the manager) as using 8 CPUs,
and will be treated by the client as using 8 CPUs
(e.g. on an 8-core machine it will be the only job running).

Note 3: the "max % of CPUs" computing preference applies to
physical as well as logical CPUs

Also: bug fix: the # of threads determined by the <mem_usage_*>
plan class elements should min'd with the number determined by
<min_cpus> and <max_threads>, rather than overriding it.