Travis: restructure travis configuration
Instead of excluding unwanted jobs this now specifically includes the jobs needed. This allows to run multiple environments or languages within a single repository. This is not fully documented but was confirmed to be supported by Travis via email.

Configuration can be tested on this page: http://config.travis-ci.org/
Original idea came from this blog entry: http://axelnilsson.tech/2017/08/04/travisci-with-multi-language-repository/