Remove unused utmp variables
We also don't need a default as we already do
for for f in /etc/utmp /var/adm/utmp /var/run/utmp; during build.
And the default that we used /etc/utmp is deprecated everywhere I
could find anyway.

In the event a UTMP_LOCATION doesn't exist, I think it would more
sense just to disable utmp in the build.