db_dump changes - use consent_type table to dump user and host stats based on consent.
User and host table dump modified - use consent table to determine if user has consented to stats. exports.
Look up STATSEXPORT consent type and determine if enabled; if so, change the SQL statement to use consent table. Only if a user has consented to statistics exports is that user's data returned and exported.