client: fix GPU detection with --detach_console
GPU detection launches a copy of the client based on the contents of
argv[0]. If argv[0] doesn't include path to the client and the current
directory is not BOINC program directory then the GPU detection fails.
Hard coding the client name in --detach_console code only works because
of the extra work CreateProcess() does to find the executable. Launching
GPU detection uses CreateProcess() in a way that it doesn't try to find
the executable.

Fix this by using the full path to the client executable in
--detach_console code. This in turn makes the full path available to GPU
detection code. This also takes care of the case of the client
executable not being named boinc.exe.