client: use full path to client executable in GPU detection
GPU detection tries to launch a copy of the client using argv[0] but
argv[0] doesn't always include path. If path is not included and current
directory is not the same as BOINC program directory then the GPU
detection fails.

Use full path to the client executable instead if possible. Quote the
path on Windows to handle paths with spaces.